Fix false fallback when client joins an existing mcast group with high traffic.
Only take action upon loss rate when the deques are at least half full.

When a client joins a multicast group with heavy traffic going on, it will lose
a lot of packets in the very beginning because there is a considerable time
amount between it's multicast socket creation and the first read. Thus, the socket
buffer is likely to overflow in this start situation, resulting in packet loss.