Android: implement colour mode selection for native connections
As it stands now, this can only be done on connection setup, as
changing pixel format while processing server messages often leads to
a corrupted stream (which the native implementation does not like and
bails out). The correct solution would be to make use of
https://github.com/rfbproto/rfbproto/blob/master/rfbproto.rst#clientfence

Also, we only support 24 and 15 bit colour depths, as the GL canvas is
not able to digest any other format _directly_, Android does not have
the GL_UNSIGNED_BYTE_3_3_2 packed pixel type.

Closes #77