Android: VncCanvas: remove scrollToAbsolute()
scrollToAbsolute() scrolls the canvas view itself to a certain
position and thus is a remnant of the previous display engine:
with the current OpenGL drawing, we adjust the texture inside
the view and need not scroll the view to any position.

scrollToAbsolute() did not do any harm on Android < 12, but
caused white artifacts on Android 12, probably due to some
optimisation on the OS side.

Thus, remove scrollToAbsolute().

Closes #183